backend/requirements.txt
amqp==5.3.1
asgiref==3.11.0
billiard==4.2.4
celery==5.6.0
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
dj-database-url==2.1.0
Django==4.2.8
django-cors-headers==4.9.0
djangorestframework==3.16.1
exceptiongroup==1.3.1
gunicorn==21.2.0
kombu==5.6.1
packaging==25.0
prompt_toolkit==3.0.52
psycopg2-binary==2.9.11
python-dateutil==2.8.2
python-decouple==3.8
redis==7.1.0
six==1.17.0
sqlparse==0.5.5
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
vine==5.1.0
wcwidth==0.2.14
whitenoise==6.6.0
